% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Standard codon table for centraldog}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table
}
\description{
This codon table can be used to replicate the process of changing DNA to RNA and vise versa aka central dogma of molecular biology
}
\keyword{datasets}
