% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DNA.R
\name{create_DNA}
\alias{create_DNA}
\title{Create a random DNA sequence of a given length.}
\usage{
create_DNA(n)
}
\arguments{
\item{n}{Length of DNA sequence to be generated.}
}
\value{
Returns a randomly generated DNA sequence of a given length.
}
\description{
Create a random DNA sequence of a given length.
}
