% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{Turn a DNA sequence to an RNA sequence (exchange T to U).}
\usage{
DNA_to_RNA(sequence)
}
\arguments{
\item{sequence}{string of a DNA sequence.}
}
\value{
Returns an RNA sequence from a DNA sequence.
}
\description{
Turn a DNA sequence to an RNA sequence (exchange T to U).
}
