% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_codons.R
\name{find_codons}
\alias{find_codons}
\title{Find codons from an RNA sequence}
\usage{
find_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{RNA sequence from which the codons will be found.}

\item{start}{Reading frame start position default set to 1.}
}
\description{
Find codons from an RNA sequence
}
