% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_codons.R
\name{translate_codons}
\alias{translate_codons}
\title{Translate codons into amino acids}
\usage{
translate_codons(codons)
}
\arguments{
\item{codons}{}
}
\description{
Translate codons into amino acids
}
